requirements.txt
streamlit_option_menu
streamlit-modal